Improve performance
By only registering (and caching) partials the first time that
`renderView` is called.

In development it can be helpful to re-register partials on each render,
so that the render reflects any changes you have made to the partials.
This behaviour can be retained by requesting re-registration in your
`renderView` call i.e.

```
handlebars.renderView("view_name",  {}, "layout_name", true)
```

A user of this package can choose to set this final parameter based on
the environment their code is running in.

Make lazy